xarray: Add XArray load operation
The xa_load function brings with it a lot of infrastructure; xa_empty(),
xa_is_err(), and large chunks of the XArray advanced API that are used
to implement xa_load.

As the test-suite demonstrates, it is possible to use the XArray functions
on a radix tree.  The radix tree functions depend on the GFP flags being
stored in the root of the tree, so it's not possible to use the radix
tree functions on an XArray.

Signed-off-by: Matthew Wilcox <willy@infradead.org>
(cherry picked from commit ad3d6c7263e368abdc151e1cc13dc78aa39cc7a7)

Conflicts:
	lib/Kconfig.debug
	lib/Makefile
	tools/testing/radix-tree/.gitignore